Fixed error loading .tif files
In motion_correction.py, the tiffile's imread() function uses a keyword argument `key` that indexes into the pages of a multi-page tif. Not all tif's are formatted as multi-page so this fails on perfectly valid tif files (e.g. those produced by saving from ImageJ). The workaround is to just use imread() without the `key` parameter which returns a Frames x Height x Width numpy array of the movie and then use np.take() function to select the frames as originally intended with the `key` parameter.