Fix bug with scaling in notebook
With both size dimensions being variable the image enlargened uncontrollably when using the slider of neurons for some reason. Fixing the width and varying the height will make sure that the image on the left and spikes on the right are both visible without having to scroll left or right and at the same time keep the right aspect ratio. Also d1 and d2 seem to be misnamed as d2 is constantly used for width.