Demo notebook for seeded CNMF
This notebook demonstrates how to use seeded CNMF. As discussed in the hackathon, the recommended way is to load individual mask files for each neuron. This makes overlapping ROIs possible and transforming the data into csc_matrix format easier. Because I already wrote the code, I left the watershed processing for separating combined ROIs in, only marked it as 'not recommended'.

To do:
Use a better dataset with nice, premade mask files that the user can download.