Add thorough description of CaImAn parameters
A thorough description of most CaImAn parameters that might be adapted by the user. Additionally, short descriptions of the params object structure. I chose to explain these parameters because I found them the least intuitive to understand, but they are still imporant for a good CaImAn performance.

I also added a detailed description of how the component evaluation works. The explanation is based on the one in the paper, but I couldn't find it in this form in the docs or the wiki (yet).

To do:
For now, I added these additions to the Getting_Started section of the docs. It might be good to add them also to the github wiki (e.g. under the "CaImAn Info and Tutorials" section as a "Parameters" page. Generally, I made the experience that the github wiki is more accessible to first-time users than the docs.