update estimates.py -- fix params in contour plotters
For Issue 465 parameter labels were added to contour plots (https://github.com/flatironinstitute/CaImAn/issues/465). However, instead of the intended `min_SNR` (which is the actual threshold) it displays `SNR_lowest`. This PR fixes that issue.

Note I have not run tests on this, but only run it locally to make sure it works as expected.